requirements.txt
django
requests
lxml.html
